kernel: message does not execute correct put front behavior
Signed-off-by: TaiJu Wu <tjwu1217@gmail.com>

This test covers the behavior of putting a message into the message queue when it is full.
In theory, it should pass, but in reality, this test fails.
The reason is that when the buffer is full, Thread A gets pended (blocked).
If Thread B later calls the get function, it will unpend Thread A, allowing it to resume and put the message into the queue.
In this situation, we need to know whether Thread A should continue with put to front or put to end.